services/generation/requirements.txt
fastapi
uvicorn
python-docx
pydantic
Pillow
reportlab
